single side PCB mainly used to connect everything
The PCB design is single sided, has no SMD parts and few parts. Hence it is very straightforward to make it at home both with a CNC or chemicals. The main purpose of the PCB is to speed up the assembly but it is better to choose terminals that guarantee contact continuity over time since this is an outdoor application partially exposed to weather conditions.